Add files via upload
About this Course
In the second course of the Practical Data Science Specialization, I learnt to automate a natural language processing task by building an end-to-end machine learning pipeline using Hugging Face’s highly-optimized implementation of the state-of-the-art BERT algorithm with Amazon SageMaker Pipelines. My pipeline first transformed the dataset into BERT-readable features and store the features in the Amazon SageMaker Feature Store. It then fine-tuned a text classification model to the dataset using a Hugging Face pre-trained model, which has learned to understand the human language from millions of Wikipedia documents. Finally, my pipeline evaluated the model’s accuracy and only deploy the model if the accuracy exceeded a given threshold.

Practical data science geared me towards handling massive datasets that do not fit in my local hardware and could originate from multiple sources. One of the biggest benefits of developing and running data science projects in the cloud is the agility and elasticity that the cloud offers to scale up and out at a minimum cost.

The Practical Data Science Specialization helped me develop the practical skills to effectively deploy my data science projects and overcome challenges at each step of the ML workflow using Amazon SageMaker. This Specialization was designed for data-focused developers, scientists, and analysts familiar with the Python and SQL programming languages and learnt howw to build, train, and deploy scalable, end-to-end ML pipelines - both automated and human-in-the-loop - in the AWS cloud.